Head Detection API

WormFeatureDetector.find_head — Function

Finds the tip of the nose of the worm in each time point, and warns of bad time points. Uses a series of blob-approximations of the worm with different sensitivities, by using local convex hull. The convex hulls should be set up in increasing order (so the last convex hull is the most generous). The difference between the first two convex hulls is used to determine the direction of the worm's head. The third convex hull is used to find the tip of the worm's head.

Arguments

centroids: the locations of the neuron centroids

imsize: the image size

Optional keyword arguments

tf (default [10,10,30]): threshold for required neuron density for convex hull i is (number of centroids) / tf[i]

max_d (default [30,50,50]): the maximum distance for a neuron to be counted as part of convex hull i is max_d[i]

hd_threshold::Integer (default 100): if convex hulls 2 and 3 give head locations farther apart than this many pixels, set error flag.

vc_threshold::Integer (default 300): if convex hulls 2 and 3 give tail locations farther apart than this many pixels, set error flag.

num_centroids_threshold::Integer (default 90): if there are fewer than this many centroids, set error flag.

edge_threshold::Integer (default 5): if the boundary of the worm is closer than this to the edge of the frame, set error flag.

manual_override: In case the algorithm finds the worm's ventral cord instead of its head, set this variable to a list of all time points where

the algorithm was wrong.

Outputs a tuple (head_pos, q_flag, crop_x, crop_y, crop_z, theta, centroid)

head_pos: position of the worm's head.

q_flag: array of error flags (empty means no errors were detected)

crop_x, crop_y, crop_z: cropping parameters

theta: amount by which to rotate the image to align it in the x-y plane

centroid: centroid of the worm

Finds the tip of the nose of the worm in each time point, and warns of bad time points. Uses a series of blob-approximations of the worm with different sensitivities, by using local convex hull. The convex hulls should be set up in increasing order (so the last convex hull is the most generous). The difference between the first two convex hulls is used to determine the direction of the worm's head. The third convex hull is used to find the tip of the worm's head.

Arguments:

param_path::Dict: Dictionary of paths including the keys:

path_head_pos: Path to head position output file

path_dir_mhd_crop: Path to MHD input files

path_dir_centroid: Path to centroid input files

param::Dict: Dictionary of parameter settings including the keys:

head_threshold: threshold for required neuron density for convex hull i is (number of centroids) / param["head_threshold"][i]

head_max_distance: the maximum distance for a neuron to be counted as part of convex hull i is max_d[i]

head_err_threshold: if convex hulls 2 and 3 give head locations farther apart than this many pixels, set error flag.

head_vc_err_threshold: if convex hulls 2 and 3 give tail locations farther apart than this many pixels, set error flag.

head_edge_err_threshold: if the boundary of the worm is closer than this to the edge of the frame, set error flag.

t_range: The time points to compute head location

f_basename::Function: Function that takes as input a time point and a channel and gives the base name of the corresponding MHD file.

manual_override (optional): In case the algorithm finds the worm's ventral cord instead of its head, set this variable to a list of all time points where the algorithm was wrong.

WormFeatureDetector.in_conv_hull — Function

Computes whether a point is in the local convex hull of a collection of 2D points.

Arguments:

point: a point in 2D space.

centroids: a cloud of points in 2D space.

max_d::Real: the farthest a point in centroids can be from point and still count for the convex hull