HSN Feature Detection API

WormFeatureDetector.find_gut_granules — Function

Finds, returns, and outputs gut granule masks for a set of images. The masks filter out the gut granules, so 1 is not a gut granule, and 0 is a gut granule.

Arguments

path::String: working directory path; all other directory inputs are relative to this

names::Array{String,1}: filenames of images to process

threshold::Real: pixel intensity brightness value. Pixels below this intensity are excluded

density::Real: density of nearby pixels that must meet the threshold for the original pixel to be counted

radius: distances in each dimension (in pixels) away from the original pixel that are counted as nearby. For example, radius = [3, 2, 1] would allow a distance of three pixels in the x-direction, two pixels in the y-direction, and one pixel in the z-direction.

Optional keyword arguments

nrrd::String: path to NRRD directory, where the image will be found. Default "NRRD".

out::String: path to output directory to store the masks. If left blank (default), images will not be written.

WormFeatureDetector.find_hsn — Function

Finds location of the HSN soma in a frame. First threshold to remove densely-packed regions that might correspond to gut fluorescence or neuropil (by excluding too-dense regions), then threshold again to ensure high local density (by excluding not-dense regions). If multiple regions are still included, the larger region is chosen. Can optionally choose to output data to a file, for use with heuristics.

Arguments:

path::String: working directory path; all other directory inputs are relative to this

frames: frames of images to process

dir_nrrd::String: path to NRRD directory, where the image will be found.

img_prefix::String: image prefix not including the timestamp. It is assumed that each frame's filename will be, eg, img_prefix_t0123_ch2.nrrd for frame 123 with channel=2.

channel::Integer: channel being used.

threshold_outer::Real: pixel intensity brightness value. Pixels below this intensity are excluded

density_outer::Real: density of nearby pixels that must meet the outer threshold for the original pixel to NOT be counted

radius_outer: distances in each dimension (in pixels) away from the original pixel that are counted as nearby.

For example, radius = [3, 2, 1] would allow a distance of three pixels in the x-direction, two pixels in the y-direction, and one pixel in the z-direction.

threshold_inner::Real: pixel intensity brightness value. Pixels below this intensity are excluded

density_inner::Real: density of nearby pixels that must meet the inner threshold for the original pixel to be counted

radius_inner: distances in each dimension (in pixels) away from the original pixel that are counted as nearby.

For example, radius = [3, 2, 1] would allow a distance of three pixels in the x-direction, two pixels in the y-direction, and one pixel in the z-direction.

radius_detection: If multiple locations remain as possible HSN locations after both thresholding steps, the location with the most other such points within radius_detection of it is chosen.

Optional keyword arguments

outfile::String: path to HSN output file. If left blank (default), no output will be written.

WormFeatureDetector.find_nerve_ring — Function

Finds location of the nerve ring in a frame. Can optionally output nerve ring locations to a file for use with heuristics.

Arguments:

path::String: working directory path; all other directory inputs are relative to this

frames: frames of images to process

dir_nrrd::String: path to NRRD directory, where the image will be found.

img_prefix::String: image prefix not including the timestamp. It is assumed that each frame's filename will be, eg, img_prefix_t0123_ch2.nrrd for frame 123 with channel=2.

channel::Integer: channel being used.

threshold::Real: pixel intensity brightness value. Pixels below this intensity are excluded

region: region of the image that will be searched for the nerve ring. Generically, you should try to include the nerve ring and exclude any other regions (such as gut granules, HSN soma, etc)

radius: The location with the most other points that meet the threshold within radius of it is chosen as the nerve ring location.

Optional keyword arguments

outfile::String: path to nerve ring output file. If left blank (default), no output will be written.

WormFeatureDetector.elastix_difficulty_hsn_nr — Function

Computes registration difficulty between two frames based on the HSN and nerve ring locations

Arguments:

rootpath::String: working directory path; all other directory inputs are relative to this

frame1::Integer: first frame

frame2::Integer: second frame

hsn_location_file::String: path to file containing HSN locations

nr_location_file::String: path to file containing nerve ring locations

Heuristic parameters (optional):

nr_weight::Real: weight of nerve ring location relative to HSN location. Default 1.